Move context out of index (react-refresh/only-export-components)
