+++ lib/components/Notifications/Inbox.tsx
@@ -2,12 +2,13 @@ import { Empty, List } from 'antd';
 import { InboxHeader, InboxHeaderProps } from './InboxHeader';
 import VirtualList from 'rc-virtual-list';
 import { Notification } from './Notification';
-import { NotificationAPIContext } from '../Provider';
+
 import { useContext } from 'react';
 import { NotificationPopupProps } from './NotificationPopup';
 import { Liquid } from 'liquidjs';
 import { InAppNotification } from '@notificationapi/core/dist/interfaces';
 import { Filter, ImageShape, Pagination } from './interface';
+import { NotificationAPIContext } from '../Provider/context';
 
 export type InboxProps = {
   pagination: keyof typeof Pagination;
