+++ lib/components/Provider/index.tsx
@@ -1,6 +1,5 @@
 import {
   PropsWithChildren,
-  createContext,
   useCallback,
   useContext,
   useEffect,
@@ -22,47 +21,7 @@ import {
   DeliveryOptionsForEmail,
   DeliveryOptionsForInappWeb
 } from '@notificationapi/core/dist/interfaces';
-
-export type Context = {
-  notifications?: InAppNotification[];
-  preferences?: GetPreferencesResponse;
-  userAccountMetaData?: { userAccountMetadata: UserAccountMetadata };
-  webPushOptInMessage?: 'AUTOMATIC' | boolean;
-  loadNotifications: (initial?: boolean) => void;
-  markAsOpened: () => void;
-  markAsArchived: (ids: string[] | 'ALL') => void;
-  markAsUnarchived: (ids: string[] | 'ALL') => void;
-  markAsClicked: (ids: string[]) => void;
-  updateDelivery: (
-    notificationId: string,
-    channel: Channels,
-    delivery:
-      | DeliveryOptionsForEmail
-      | DeliveryOptionsForInappWeb
-      | BaseDeliveryOptions,
-    subNotificationId?: string
-  ) => void;
-  updateDeliveries: (
-    params: {
-      notificationId: string;
-      channel: Channels;
-      delivery:
-        | DeliveryOptionsForEmail
-        | DeliveryOptionsForInappWeb
-        | BaseDeliveryOptions;
-      subNotificationId?: string;
-    }[]
-  ) => void;
-  getClient: () => typeof NotificationAPIClientSDK;
-  setWebPushOptInMessage: React.Dispatch<
-    React.SetStateAction<'AUTOMATIC' | boolean>
-  >;
-  setWebPushOptIn: React.Dispatch<React.SetStateAction<boolean>>;
-};
-
-export const NotificationAPIContext = createContext<Context | undefined>(
-  undefined
-);
+import { Context, NotificationAPIContext } from './context';
 
 type Props = (
   | {
